+++ AddOns/HandyNotes_TheWarWithin/core/hook.lua
@@ -71,7 +71,8 @@ end
 local Vignette = Class('Vignette', Hook, {
     type = 'vignette',
     rewardsSpaceBefore = false,
-    rewardsSpaceAfter = true
+    rewardsSpaceAfter = true,
+    group = ns.groups.MISC
 })
 
 function Vignette:Initialize(attrs)
@@ -168,12 +169,23 @@ local function HookAllPOIS()
         local poi = ns.hooks['delve'][self.poiInfo.areaPoiID]
         if not poi then return end
 
-        local hookInfo = Hook({
-            rewardsSpaceBefore = true,
-            rewardsSpaceAfter = false,
-            rewards = poi.rewards,
-            group = poi.group
-        })
+        local hookInfo
+
+        if (self.poiInfo.description == _G.DELVE_LABEL) then
+            hookInfo = Hook({
+                rewardsSpaceBefore = true,
+                rewardsSpaceAfter = false,
+                rewards = poi.rewards,
+                group = poi.group
+            })
+        else
+            hookInfo = Hook({
+                rewardsSpaceBefore = false,
+                rewardsSpaceAfter = true,
+                rewards = poi.rewards,
+                group = poi.group
+            })
+        end
 
         renderTooltip(self, hookInfo)
     end)
@@ -286,9 +298,9 @@ local function HookAllPOIS()
         local vignetteGUID = self.vignetteGUID
         local pos = C_VignetteInfo.GetVignettePosition(vignetteGUID, mapID)
         local coordinates = HandyNotes:getCoord(pos.x, pos.y)
-        local node = ns.maps[mapID].nodes[coordinates]
-        hookInfo.note = node.note
-        hookInfo.rewards = node.rewards
+        local node = ns.maps[mapID] and ns.maps[mapID].nodes[coordinates]
+        if node and node.note then hookInfo.note = node.note end
+        if node and node.rewards then hookInfo.rewards = node.rewards end
         renderTooltip(self, hookInfo)
     end)
 end