+++ AddOns/HandyNotes_TheWarWithin/localization/zhTW.lua
@@ -25,6 +25,9 @@ L['options_icons_flight_master_desc'] = nil
 L['options_icons_worldsoul_memories'] = nil
 L['options_icons_worldsoul_memories_desc'] = nil
 
+L['last_architect_note'] = nil
+L['neighborhood_riddle'] = '寶藏位置'
+
 -------------------------------------------------------------------------------
 -------------------------------- ISLE OF DORN ---------------------------------
 -------------------------------------------------------------------------------
@@ -675,3 +678,15 @@ L['ancient_coffer_note'] = nil
 
 L['sthaarbs_note'] = nil
 L['heka_tamos_note'] = nil
+
+L['translocated_gorger_note'] = nil
+
+L['zo_turu_note_start'] = nil
+L['zo_turu_note_end'] = nil
+L['cartel_ba_dead_drop_label'] = nil
+L['cartel_ba_dead_drop_location'] = nil
+L['cartel_om_dead_drop_label'] = nil
+L['cartel_om_dead_drop_location'] = nil
+L['cartel_zo_dead_drop_label'] = nil
+L['cartel_zo_dead_drop_location'] = nil
+L['cartel_dead_drop_warning'] = nil