+++ AddOns/HandyNotes_TheWarWithin/core/CHANGELOG.md
@@ -1,3 +1,30 @@
+# v143
+
+* Updated tooltip item function to avoid potential taint issue.
+
+# v137
+
+* Updated primary TOC interface version to 12.0.1
+
+# v132
+
+* Update Krowi's world map button.
+
+# v131
+
+* Updated all plugins for compatibility with 12.0.0 UI changes.
+
+# v129
+
+* Updated Ace3 Libraries
+* Moved profession treasure functionality to core
+* Added new decor-related options and templates
+* Updated core localizations: deDE, enUS, esES, esMX, frFR, koKR, ruRU, zhCN, zhTW
+
+# v123
+
+* Fixed lua error in pin tooltip [[#503]](https://github.com/zarillion/handynotes-plugins/issues/503)
+
 # v118
 
 * Added a new approach to hooking POI vignettes.